backend/requirements.txt
Django==5.1.2
django-cors-headers==4.3.1
requests==2.32.3
python-dateutil==2.8.2
icalendar==5.0.11
